requirements.txt
pdfplumber>=0.10.0
python-docx>=0.8.0
markdown>=3.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.37.0
safetensors>=0.3.0
llama-cpp-python>=0.2.0
click>=8.0
pyyaml>=6.0
tqdm>=4.65